requirements.txt
openpyxl==3.0.10
pandas==1.5.3
pandas-datareader==0.9.0
pyarrow==10.0.1
PyYAML==6.0.1
requests==2.22.0
wget==3.2
matplotlib==3.6.3
statsmodels==0.13.5
numpy==1.24.4
scipy==1.9.3
graphviz==0.20
numba==0.57.1
sequence-jacobian==1.0.0